Add a slideshow for git
This was prepared for a team guild meeting. Figured I might as well
reused this framework and write all my slides in markdown since markdown
is the best.

All the framework (reveal.js and the like) is included in the shared
folder so that different slide topics can be created with minimal
duplication.