Create index.html page; enable latest reveal.js
The latest revel.js is many versions newer than the version that was
hard-included in this repo. Now I've got a submodule and can start using
the latest version with newer slides. Eventually converting older slides
over as well. Also, this integrates a theme override that lines up more
with the theme on gidgidonihah.com

Also added an index.html for indexing the public slides, rather than the
markdown doc. This will allow me to host it on a simple nginx page at
slides.gidgidonihah.com.